Cleanup IE Makefile
We remove the following targets from the top-level IE Makefile
- load-secrets:
- vault-init:
- argosecret:

Reason is that they are identical to what's in common/Makefile
We can probably also remove the upgrade target, but since it has a
dependency on load-secrets (which the common/Makefile upgrade target
does not have), I am leaving it out now.

This reduces the Makefile complexity a little bit.