Do not bail out when KUBECONFIG is unset
We first check for ~/.kube/config and if it also does not exist
then we bail out.

There should be no need to set anything else in the playbook as
~/.kube/config is automatically read by the ansible kubernetes.core
collection [1]

Note: This change is for the ansible in IE only. We will drop this once
we move common/ to ansible and rebase IE to use those common/ roles.

[1] https://github.com/ansible-collections/kubernetes.core/blob/main/docs/kubernetes.core.kubectl_connection.rst