Remove SECRETS from the initial helm chart args passing
We're using vault + the load-secrets target to inject secrets in it.
Let's not pass the secret values via helm anywhere as this really
defeats the whole purpose of uploading the secrets out of band.

This is the same commit as af5bf8d26538aaa59d2f750112106471ab73d30c we
did in common.