Update to AMC policy: factory-data-lake-secret-policy.yaml
This update is to handle propagation of the Kafka Cluster CA to multiple
edge clusters.  The propagation is done via an ACM policy.  We can define
multiple clusters, with different labels, under the managedClusterGroups
section in the values-hub/datacenter.yaml file.  The list of edge clusters
define the target clusters where we want to deploy the Industrial Edge
Stormshift applications. The update ensures that ACM will propagate the CA for
the manuela-data-lake cluster to these clusters.