Switch to community operator for camel-k
When writing to the s3 endpoint exported by nooba, we need a feature in
camel called `forcePathStyle`, because the s3 endpoint does not work
with virtual hosts like proper S3 from AWS.

This `forcePathStyle` support showed up in camel-3.21.x. Currently
the downstream camel-k only ships version 3.18.6 which will error out
as follows:

    Failed to resolve endpoint:
    aws2-s3://anomaly-detection?accessKey=xxxxxx&forcePathStyle=true&overrideEndpoint=true&region=none&secretKey=xxxxxx&uriEndpointOverride=http%3A%2F%2Fs3.openshift-storage.svc
    due to: There are 1 parameters that couldn't be set on the endpoint.
    Check the uri if the parameters are spelt correctly and that they are
    properties of the endpoint. Unknown parameters=[{forcePathStyle=true}]

There are no newer runtimes we can use in the CamelCatalog provided by
the camel-k downstream operator currently (operator version
1.10.7+xxxx).

So for the time being let's switch to the upstream camel k operator
which ships a camel catalog that is more recent:

    oc get -n openshift-operators camelcatalogs camel-catalog-3.8.1 -o yaml |grep camel.version
    camel.version: 4.4.1

With this version we're able to set the proper S3 URI and the
application is able to write to the nooba s3 bucket correctly.

Co-Authored-By: Akos Eros <aeros@redhat.com>