Switch to traits for kafka integrations
The newly released camel-k 2.5.0 dropped support for the old
spec.configuration mount points and we need to switch to the mount
traits (https://camel.apache.org/camel-k/2.5.x/traits/mount.html)

This seems to mean that the keys inside a secret need to be under the
`application.properties` umbrella, for some java reason that is
escaping us. Amend the s3-anomaly secret creation so that everything
works.

Co-Authored-By: Akos Eros <aeros@redhat.com>