Talk to the model server on the hub for now
This is an initial pass at making a stock deployment to just work.
Basically we tell the messaging component on a factory (which represents
a production workload) to talk to the inference server in the
manuela-tst-all namespace on the hub.

This change will need more in-depth discussion in the future and we need
to clear up the exact workflow. We could decide to spin up an inference
server in the factory as well and have it load a model from the hub's
S3, but it might get a bit too onerous for a demo setup like this one.

So for the time being, let's use this approach to at least get the line
dashboard working on the factory, out of the box.

Co-Authored-By: Akos Eros <aeros@redhat.com>