Add storage overrides for different clouds
The helm variable `global.datacenter.storageClassName` is used by the
ODF chart to pick the storage class for its storageclass.
It defaults to `gp3-csi` which only exists in AWS.

Let's add three overrides for the major clouds, picking the right
storageclass for that cloud. Also add one for the AWS one just to make
things more explicit.

Reported-By: Mark Labonte <mlabonte@redhat.com>