Serialize tasks when used in the seed pipeline
mlabonte saw some racey behaviour in the `make seed` pipeline.
It seems we are exposed to some races due to the parallelization
of some tasks and the use of yq on files.

For the time being let's go back at running things serially.
We can investigate later on how to speed things up a bit more if needed.

Reported-By: Mark Labonte <mlabonte@redhat.com>
Co-Authored-By: Akos Eros <aeros@redhat.com>