Switch to docker.io/seldonio/mlserver:1.3.5-sklearn
This image is almost a 1/4th of the 1.3.5 bigger image 0.93GB vs 3.74GB
and since we use the `sklearn` model this is okay to do and alleviates
some of the pain we see due to the download size.

Tested this and the line-dashboard kept working correctly on both hub
and factory.