Export TRUSTED_CA_BUNDLE_PATH in the notebooks and bind mount the certs in the main notebook
According to https://elyra.readthedocs.io/en/stable/user_guide/pipeline-components.html?highlight=ssl
seeting TRUSTED_CA_BUNDLE_PATH is how you set the CA pem file for elyra pipelines.

But in order to do that we also need to bind mount the tls mount point
in the main jupyter so that we can actually reference that CA bundle
which contains both the official CAs and the internal cluster ones.