Make sure integrations work with camel-k > 2.5
In camel-k 2.5 it all used to work automatically and builds would push
things to the internal registry using s2i. Since that has been
deprecated we now, add a long-lived token to the camel-k-operator
service account, then we push that secret to vault and use it via
and external secret to create a proper dockercfgjson secret called
`ocp-registry` which will then be used by the IntegrationPlatform
objects.